fix: simplify log handler and resolution classes
