+++ src/handlers/observability/types.ts
@@ -1,7 +1,7 @@
 import type z from "zod";
 import type { ObservableResourceRef } from "../../core/observability";
 import type { CoreOptions } from "../../core/types";
-import type { Context, Flag, Handler } from "../../router";
+import type { Context, Flag } from "../../router";
 
 export type ResourceFlagValues<F extends readonly Flag<string, unknown>[]> = {
   [E in F[number] as E["name"]]: E extends Flag<string, infer T> ? z.infer<z.ZodType<T>> : never;
@@ -22,12 +22,3 @@ export type ResolvedObservableResource<K extends ObservableResourceRef["kind"]> 
   resource: Extract<ObservableResourceRef, { kind: K }>;
   options?: CoreOptions;
 };
-
-export interface ObservabilityHandlerFactories {
-  createLogsHandler<
-    K extends ObservableResourceRef["kind"],
-    F extends readonly Flag<string, unknown>[],
-  >(config: {
-    resource: ObservableResourceCommand<K, F>;
-  }): Handler;
-}
